+++ src/diagnostics/tsx/validators/button.ts
@@ -1,12 +1,7 @@
-import {
-  JSXOpeningElement,
-  JSXIdentifier,
-  JSXSpreadAttribute,
-  JSXAttribute,
-} from '@babel/types';
 import { messages } from '../messages';
 import { Diagnostic } from '../diagnostic';
 import { Validator } from './validator';
+import { TSXElement } from '../element';
 
 export class ButtonValidator implements Validator {
   #tags: string[] = ['button'] as const;
@@ -15,22 +10,11 @@ export class ButtonValidator implements Validator {
     return this.#tags;
   }
 
-  validate(node: JSXOpeningElement): Diagnostic[] {
-    const { loc, attributes } = node;
-
-    if (!this.hasAttribute('aria-label', attributes)) {
-      return [new Diagnostic(messages.button['aria-label'], loc)];
+  validate(node: TSXElement): Diagnostic[] {
+    if (!node.hasAttribute('aria-label')) {
+      return [new Diagnostic(messages.button['aria-label'], node.loc)];
     }
 
     return [];
-  }
-
-  private hasAttribute(
-    attribute: string | JSXIdentifier,
-    attributes: (JSXAttribute | JSXSpreadAttribute)[]
-  ): boolean {
-    return attributes.some(
-      (attr) => attr.type === 'JSXAttribute' && attr.name.name === attribute
-    );
   }
 }