+++ CHANGELOG.md
@@ -7,11 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5]- 2024-08-27
+
 ### Added
 
-- **TSX Validation:**
-- Added support for tsx files.
-- Added validation of `<img>` elements (e.g., alt attribute).
+**TSX Validation:**
+
+- Added support for .tsx files.
+- Added validation for <img> elements (e.g., checking for alt attributes).
+
+**HTML Validation:**
+
+- Added check for consecutive <div> elements.
+- Added check for consecutive <a> elements.
 
 ## [0.0.4]- 2024-08-21